% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotestthat-package.R
\docType{package}
\name{autotestthat-package}
\alias{autotestthat}
\alias{autotestthat-package}
\title{autotestthat: Easily automate your test-driven development in R}
\description{
The goal of autotestthat is to provide some easy way to
             use `testthat::auto_test_package()` functionality on
             RStudio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/autotestthat}
  \item Report bugs at \url{https://github.com/CorradoLanera/autotestthat/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

}
\keyword{internal}
